Add support for bridge bots (#99)
* Add support for bridge bots

On #UrLab, we have the bot LeBID (Le Bridge-IRC-Discord). As its name
suggests, if shares the messages between IRC and Discord. When arriving on IRC
the messages are in the form: "<DiscordUser> text that was sent on Discord".

Because of this, many LechBot handlers are not triggerd, because they expect
the text to start with the command directly (like "!motd ..."), and as such
some commands are unavailable to Discord users.

We therefore add a mechanism when a new message is received, such that if a
known bridge bot sent the message, we extract the original user and text before
processing it like any other IRC message.

Signed-off-by: iTitou <moiandme@gmail.com>

* Fix black

Signed-off-by: iTitou <moiandme@gmail.com>